chore: rename localStorage to storage
