+++ src/features/settingSaga.ts
@@ -2,7 +2,7 @@ import { PayloadAction } from '@reduxjs/toolkit'
 import { call,takeLatest } from 'redux-saga/effects'
 
 import { redirect } from '@/utils/history'
-import { localStorageUtil, STORAGE_KEY } from '@/utils/localStorage'
+import { localStorageUtil, STORAGE_KEY } from '@/utils/storage'
 
 import { userInfoActions } from './settingSlice'
 function* saveUserToken({payload}: PayloadAction<string>) {
